pyproject.toml
[tool.poetry]
name = "indexer"
version = "0.1.0"
description = "Web3 indexer with Apibara"
authors = []

[tool.poetry.scripts]
indexer = "indexer.main:cli"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
apibara = "^0.5.15"
click = "^8.1.3"
pymongo = {extras = ["srv"], version = "^4.1.1"}
strawberry-graphql = {extras = ["debug-server"], version = "^0.133.5"}
aiohttp = "^3.8.3"
websockets = ">6.0.0"
web3 = ">4.10.0"
pre-commit = "^2.20.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"
pytest-cov = "^4.0.0"
mongomock = "^4.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
